rename(payment -> client): payment 패키지를 client 패키지 안으로 이동
