+++ src/main/java/roomescape/config/ClientConfig.java
@@ -4,14 +4,14 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.client.RestClient;
-import roomescape.payment.service.PaymentService;
+import roomescape.client.payment.PaymentClient;
 
 @Configuration
 public class ClientConfig {
 
     @Bean
-    public PaymentService paymentClient(@Value("${payment.widget-secret-key}") String widgetSecretKey) {
+    public PaymentClient paymentClient(@Value("${payment.widget-secret-key}") String widgetSecretKey) {
         RestClient restClient = RestClient.builder().build();
-        return new PaymentService(widgetSecretKey, restClient);
+        return new PaymentClient(widgetSecretKey, restClient);
     }
 }
